requirements.txt
numpy>=1.21
Pillow>=9.0
opencv-python>=4.5S
tqdm>=4.64
ipython>=8.0
